Display parameters in describe binding output
When no parameters are defined, the entire section is not displayed
