+++ cmd/svcat/output/instance.go
@@ -17,12 +17,10 @@ limitations under the License.
 package output
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
 
 	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
-	"k8s.io/apimachinery/pkg/runtime"
 )
 
 func getInstanceStatusCondition(status v1beta1.ServiceInstanceStatus) v1beta1.ServiceInstanceCondition {
@@ -119,21 +117,5 @@ func WriteInstanceDetails(w io.Writer, instance *v1beta1.ServiceInstance) {
 		writeParameters(w, instance.Status.ExternalProperties.Parameters)
 	} else if instance.Spec.Parameters != nil {
 		writeParameters(w, instance.Spec.Parameters)
-	}
-}
-
-func writeParameters(w io.Writer, parameters *runtime.RawExtension) {
-	if parameters == nil {
-		return
-	}
-
-	fmt.Fprintln(w, "\nParameters:")
-	var params map[string]interface{}
-	err := json.Unmarshal(parameters.Raw, &params)
-	if err != nil {
-		// If it isn't formatted in json, just show the string representation of what is present
-		fmt.Fprintln(w, string(parameters.Raw))
-	} else {
-		writeYAML(w, params, 2)
 	}
 }
