+++ cmd/svcat/testdata/responses/namespaces/test-ns/servicebindings/ups-binding.json
@@ -34,7 +34,7 @@
     "asyncOpInProgress": false,
     "reconciledGeneration": 1,
     "externalProperties": {
-      "parameters": {},
+      "parameters": {"param1": "value1", "paramset": {"ps1":1, "ps2": "two"}},
       "parameterChecksum": "44136fa355b3678a1146ad16f7e8649e94fb4fc21fe77e8310c060f61caaff8a"
     },
     "orphanMitigationInProgress": false,
